backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.1
pydantic-settings>=2.3.4
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
redis>=5.0.4
celery>=5.3.6
PyPDF2>=3.0.1
python-docx>=1.1.0
Pillow>=10.3.0
pytesseract>=0.3.10
python-dateutil>=2.9.0.post0
pytz>=2024.1
spacy>=3.7.4
boto3>=1.34.150
python-multipart>=0.0.9
httpx>=0.27.0
pydantic-ai>=0.0.10
openai>=1.30.0
pdf2image>=1.17.0

pytest>=8.2.1
pytest-asyncio>=0.23.7
